requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.3
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
multidict==6.2.0
mutagen==1.47.0
propcache==0.3.0
pycparser==2.22
pycryptodomex==3.22.0
PyNaCl==1.5.0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
validators==0.35.0
websockets==15.0.1
yarl==1.18.3
yt-dlp==2025.6.9.234059.dev0
